+++ backend/utils/math.go
@@ -3,7 +3,7 @@ package utils
 import "math"
 
 func GetDiff(fromX, fromZ, toX, toZ float64) (float64, float64) {
-	return float64(toX - fromX), float64(toZ - fromZ)
+	return toX - fromX, toZ - fromZ
 }
 
 func GetDistance(a, b float64) float64 {
